Fix double dispose of GCHandle in BrowserWebSocket
MemoryHandle is a struct, so passing it into an IDisposable parameter
creates a box. This causes two copies of GCHandle to exist which are
no longer synchronized and can be double disposed.

This happened when passing pinBuffer into CancellationHelper in
ReceiveAsyncCore and SendAsyncCore. In both cases the caller does the
Dispose itself inside a finally, so there's no need to pass the
pinBuffer to CancellationHelper at all.